requirements.txt
python>=2.7
biopython>=1.6
